+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 - COU level API User
 - Support VO Members REST API v2
+- Support Syncing Assurance and Cert Models from Job Shell
 
 ## [3.4.1-rciam] - 2022-02-11